resources: move various get<Foo> and getOrCreate<Foo> funcs to new file
Because smbshare.go is getting huge.
